# application/static/swagger.yaml
swagger: '2.0'
info:
  title: "Mechanic Shop API"
  description: "A comprehensive API for managing a mechanic shop including customers, mechanics, service tickets, and inventory"
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: "JWT Authorization header using the Bearer scheme. Example: 'Bearer {token}'"
paths:
...
  /service-tickets:
    post:
      tags:
        - Service Tickets
      summary: "Create a new service ticket"
      description: "Create a new service ticket for a customer's vehicle"
      parameters:
        - in: "body"
          name: "body"
          description: "Service ticket information"
          required: true
          schema:
            $ref: "#/definitions/CreateServiceTicketPayload"
      responses:
        201:
          description: "Service ticket created successfully"
          schema:
            $ref: "#/definitions/ServiceTicketResponse"
          examples:
            application/json:
              id: 1
              customer_id: 1
              vehicle_year: 2020
              vehicle_make: "Toyota"
              vehicle_model: "Camry"
              vehicle_vin: "1HGBH41JXMN109186"
              description: "Oil change and tire rotation needed"
              estimated_cost: 150.00
              status: "Open"
        400:
          description: "Invalid input"
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: "Customer not found"
          schema:
            $ref: "#/definitions/ErrorResponse"

    get:
      tags:
        - Service Tickets
      summary: "Returns all Service Tickets"
      description: "Endpoint to retrieve a list of all service tickets."
      responses:
        200:
          description: "Retrieved Service Tickets Successfully"
          schema:
            $ref: "#/definitions/AllServiceTickets"
  /service-tickets/{ticket_id}:
    get:
      tags:
        - Service Tickets
      summary: "Get a specific service ticket by ID"
      description: "Retrieve detailed information about a specific service ticket"
      parameters:
        - in: "path"
          name: "ticket_id"
          type: "integer"
          required: true
          description: "ID of the service ticket to retrieve"
      responses:
        200:
          description: "Service ticket retrieved successfully"
          schema:
            $ref: "#/definitions/ServiceTicketResponse"
        404:
          description: "Service ticket not found"
          schema:
            $ref: "#/definitions/ErrorResponse"

    put:
      tags:
        - Service Tickets
      summary: "Update a service ticket"
      description: "Update an existing service ticket's information"
      parameters:
        - in: "path"
          name: "ticket_id"
          type: "integer"
          required: true
          description: "ID of the service ticket to update"
        - in: "body"
          name: "body"
          description: "Updated service ticket information"
          required: true
          schema:
            $ref: "#/definitions/UpdateServiceTicketPayload"
      responses:
        200:
          description: "Service ticket updated successfully"
          schema:
            $ref: "#/definitions/ServiceTicketResponse"
        400:
          description: "Invalid input"
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: "Service ticket not found"
          schema:
            $ref: "#/definitions/ErrorResponse"

    delete:
      tags:
        - Service Tickets
      summary: "Delete a service ticket"
      description: "Remove a service ticket from the system"
      parameters:
        - in: "path"
          name: "ticket_id"
          type: "integer"
          required: true
          description: "ID of the service ticket to delete"
      responses:
        200:
          description: "Service ticket deleted successfully"
          schema:
            $ref: "#/definitions/DeleteResponse"
        404:
          description: "Service ticket not found"
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  CreateServiceTicketPayload:
    type: "object"
    properties:
      customer_id:
        type: "integer"
        example: 1
      vehicle_year:
        type: "integer"
        example: 2020
      vehicle_make:
        type: "string"
        example: "Toyota"
      vehicle_model:
        type: "string"
        example: "Camry"
      vehicle_vin:
        type: "string"
        example: "1HGBH41JXMN109186"
      description:
        type: "string"
        example: "Oil change and tire rotation needed"
      estimated_cost:
        type: "number"
        format: "float"
        example: 150.00
      status:
        type: "string"
        example: "Open"
    required:
      - customer_id
      - description
  UpdateServiceTicketPayload:
    type: "object"
    properties:
      vehicle_year:
        type: "integer"
        example: 2020
      vehicle_make:
        type: "string"
        example: "Toyota"
      vehicle_model:
        type: "string"
        example: "Camry"
      vehicle_vin:
        type: "string"
        example: "1HGBH41JXMN109186"
      description:
        type: "string"
        example: "Oil change, tire rotation, and brake inspection"
      estimated_cost:
        type: "number"
        format: "float"
        example: 250.00
      actual_cost:
        type: "number"
        format: "float"
        example: 245.50
      status:
        type: "string"
        example: "In Progress"
  ServiceTicketResponse:
    type: "object"
    properties:
      id:
        type: "integer"
        example: 1
      customer_id:
        type: "integer"
        example: 1
      mechanic_id:
        type: "integer"
        example: 1
      vehicle_year:
        type: "integer"
        example: 2020
      vehicle_make:
        type: "string"
        example: "Toyota"
      vehicle_model:
        type: "string"
        example: "Camry"
      vehicle_vin:
        type: "string"
        example: "1HGBH41JXMN109186"
      description:
        type: "string"
        example: "Oil change and tire rotation needed"
      estimated_cost:
        type: "number"
        example: 150.00
      actual_cost:
        type: "number"
        example: 145.00
      status:
        type: "string"
        example: "Open"
  AllServiceTickets:
    type: "array"
    items:
      type: "object"
      properties:
        id:
          type: "integer"
          example: 1
        customer_id:
          type: "integer"
          example: 1
        mechanic_id:
          type: "integer"
          example: 1
        vehicle_year:
          type: "integer"
          example: 2020
        vehicle_make:
          type: "string"
          example: "Toyota"
        vehicle_model:
          type: "string"
          example: "Camry"
        description:
          type: "string"
          example: "Oil change and tire rotation needed"
        status:
          type: "string"
          example: "Open"

  # Inventory Definitions
...
  DeleteResponse:
    type: "object"
    properties:
      message:
        type: "string"
        example: "Successfully deleted"
  ErrorResponse:
    type: "object"
    properties:
      error:
        type: "string"
        example: "An error occurred"
      errors:
        type: "object"